--- Code/OfRectanglesAndSquares.Java/src/OfRectanglesAndSquares/ISquare.java
@@ -1,12 +1,0 @@
-package OfRectanglesAndSquares;
-
-/**
- * Generic interface to a square
- */
-public interface ISquare {
-
-    public void setSide(double side);
-    
-    public double getSide();
-
-}
