+++ Code/OfRectanglesAndSquares.Java/src/interfaces/ISquare.java
@@ -1,0 +1,20 @@
+package interfaces;
+
+/**
+ * Generic interface to a square
+ */
+public interface ISquare {
+
+	/**
+	 * Sets the side of the square
+	 * @param side
+	 */
+    public void setSide(double side);
+    
+    /**
+     * returns the side of the square
+     * @return
+     */
+    public double getSide();
+
+}
